Issue gimp-ux#176: new GimpCheckExpander widget.
Core-only widget for now. The idea is a widget which has both an
enabled/checked state and an expanded/collapse state which will show a
child widget.

When enabling or disabling the main boolean value, it also triggers the
expansion of the child widget, but the opposite is not true.