app: use the new GimpCheckExpander widget for Dynamics settings.
This is a first usage mostly for testing.

This new widget doesn't use bold text to show it will expand
sub-settings, unlike the previous expanding frame (issue gimp-ux#176).
Instead it shows a small triangle on the side, as is customary for
expanders.

Furthermore the expansion state is stored as a separate property, which
means you can enable dynamics while still hiding all sub-options, hence
saving space in the tool options (issue gimp-ux#311).