widgets: Convert channel previews to u8 precision
Rather than using a hardcoded byte count check to determine
the precision of an image preview, we can retrieve the temp_buf's
GeglBuffer and then retrieve the pixel values from it using the
u8 precision version of the buffer's format.
This allows us to display the correct 8-bit preview even if the
image preview is a higher precision (or in the future, if the image is
in CMYK format).