build/windows: Check if Installer and MSIX are identical to bundle
This way, we avoid repeating the hellish situation of the past on
which the Installer scripts ruled what is distributed, which do
not make sense anymore with the advent of the bundling scripts on
CI (it breaks "deps > build/bundle > distribution" chronology).

Nowadays, it is not even sustainable with the advent of the MSIX,
which aims to have the same content as the installer and vice versa.
So, let's prevent disparity by adding a check in both dist scripts.

In short, if we want something to be included or not in the installer,
we need first to change that on 2_bundle-gimp-uni_base.py script and
change a bit on the .iss script so both packages benefit from it.