Revert "PDB: fix #14205 image-find-next-sample-point error"
This reverts commit 69894d8bbff24c85f8d98c65e0a2951ae7f8d1e2.

Let's not randomly make our API less strict on types as a workaround for
a bug. If there are specific cases where it's needed to accept 0, then
these are the cases we have to handle.

Also unlike what the commit message was saying, it's not true that this
was only affecting Script-Fu. Any change to the PDB affects by
definition the whole PDB, as well as libgimp and all bindings (over PDB
or libgimp alike). And yes, this change **was** of "real consequence".

I'll do further commits for the proper fixes to #14205.

(cherry picked from commit 2d2f1d048dc661d1bd3f84e59ae6931135e011bd)