icons: add some authorship details in COPYING.
As raised by rangelovd, most work in this legacy icon set was done by
Barbara Muraus. Looking at the git log of the SVG file noted in the
COPYING file, the only other contributor to this set seems to be Jakub
(especially at the start and end, though Barbara seems indeed to have a
lot more commits there, so it is fair to put her name prominently).