app: refactor code for 'cut' operation, allow cut for layer groups
This patch aims to reduce the code for 'cut', while enabling cutting
layer groups. Also a change in how 'have_writable' is calculated.
Layer group items are checked individually for content lock. If even
1 item has a content lock, the entire layer group cannot be cut.
If many items are selected and any 1 has content lock, none of the
items can be cut.

Issue #14074 is resolved for 'copy', this patch resolves 'cut' as well.