core: Apply filters without GUIs as NDE contextually
If someone applies a GEGL filter without a GUI like Invert
on top of a layer with existing filters, they likely expect the
filter to apply on top. Currently, we always apply filters
without GUIs destructively, which can be confusing if you
have a filter that covers the entire layer like Simplex Noise.

This patch adds a check if the layer has any visible filters,
and if so, the filter is committed non-destructively. If all
filters are currently invisible, the non-GUI filter is still
applied destructively as in 5e3047c7.