plug-ins: add convert_psd_effect_mode function...
and use it to determine the layer mode for the psd dropshadow and
solid fill effects when loaded from a descriptor.

The layer mode names (abbreviations) used in descriptors are different
from the ones Photoshop used in other (legacy) places. So we introduce
a new function here that helps converting these to an equivalent
GIMP layer mode.