layer-modes: Enable Behind for layers
We have an existing Behind layer mode, which acts as
if you're painting on the layer underneath it.
It was only available for paint tools however.
Other art programs like Krita and Photoshop allow
this blending mode to be used for layers as well.
This patch removes the restriction for this blending
mode so it can be used for layers and groups too.