UX Update 5: Fix Discard and set boundaries to path size
* In prior commits, I added custom code to
transform the layer based on the path. When
the path is discarded during rasterizing, this
caused the display not to refresh as a normal
layer. This patch adds a passthrough
function if there's no path.

* We also now get the path's boundaries
when creating and refreshing the layer,
rather than using the canvas size.
This probably still needs to be worked on.