plug-ins: Use original DDS format on export
This patch adds a parasite on load that retains
the original compression format of a DDS
texture. This parasite is then checked on export,
and if it exists, we default the compression
format to the original to reduce the chance
the user will choose the wrong format for
the game they're creating the texture for.