libgimp: fix #14641 After new metadata code, all edge builds are failing...
at devel-docs

Due to me using an enum name that only ended in one word instead of
the other enums also adding `SAVE_`, the enum values in the gir
scanner were assigned a double name, causing API breakage and a
failure to build the documentation.

So, let's conform to the other enums and change the name of our
enum to `GIMP_METADATA_SAVE_UPDATE`.