app, libgimp: Fix warning that fcinit is required since fontconfig 2.17
A recent change in Fontconfig 2.17 is that `FcInit()` became mandatory
for proper initialization. Not having that generates a warning in the
terminal:
`Fontconfig warning: using without calling FcInit()`

We fix this in the same way as GTK3 does: by calling
`pango_cairo_font_map_get_default ()`
See https://gitlab.gnome.org/GNOME/gtk/-/issues/7639

We need to add this in two places, in app for GIMP itself, and in
libgimp for plug-ins.
This adds a dependency on pangocairo for libgimp.