dialogs: make the keyboard shortcuts dialog hint more informative
This commit updates the hint in the keyboard shortcuts dialog to provide
more detailed instructions on how to edit shortcut keys. The hint now
explains that users need to click on the "Shortcut" column of the corresponding
row to type a new accelerator or press backspace to clear it.
See: https://gitlab.gnome.org/Teams/GIMP/Design/gimp-ux/-/issues/594